backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
scikit-learn
pillow
pillow-heif
rembg
numpy
sqlalchemy
onxruntime
sqlalchemy
google-generativeai